failed logins changes for unknown users, option and procedure to remove tables and settings from DB on uninstall

+++ wp-security-audit-log.php
@@ -2,9 +2,9 @@
 /*
 Plugin Name: WP Security Audit Log
 Plugin URI: http://www.wpwhitesecurity.com/wordpress-security-plugins/wp-security-audit-log/
-Description: Identify WordPress security issues before they become a problem. Keep track of everything happening on your WordPress including WordPress users activity. Similar to Windows Event Log and Linux Syslog, WP Security Audit Log generates a security alert for everything that happens on your WordPress blogs and websites. Use the Audit Log Viewer included in the plugin to see all the security alerts.
+Description: Identify WordPress security issues before they become a problem and keep track of everything happening on your WordPress, including WordPress users activity. Similar to Windows Event Log and Linux Syslog, WP Security Audit Log will generate a security alert for everything that happens on your WordPress blog or website. Use the Audit Log Viewer included in the plugin to see all the security alerts.
 Author: WP White Security
-Version: 1.3.3
+Version: 1.3.2
 Text Domain: wp-security-audit-log
 Author URI: http://www.wpwhitesecurity.com/
 License: GPL2
@@ -98,7 +98,7 @@ class WpSecurityAuditLog {
 	 * Standard singleton pattern.
 	 * WARNING! To ensure the system always works as expected, AVOID using this method.
 	 * Instead, make use of the plugin instance provided by 'wsal_init' action.
-	 * @return WpSecurityAuditLog Returns the current plugin instance.
+	 * @return \self Returns the current plugin instance.
 	 */
 	public static function GetInstance(){
 		static $instance = null;
@@ -139,6 +139,10 @@ class WpSecurityAuditLog {
 		
 		// render wsal footer
 		add_action('admin_footer', array($this, 'RenderFooter'));
+		
+		// hide plugin
+		if($this->settings->IsIncognito())
+			add_action('admin_head', array($this, 'HidePlugin'));
 	}
 	
 	/**
@@ -171,10 +175,6 @@ class WpSecurityAuditLog {
 		$s = $this->profiler->Start('WSAL Init Hook');
 		do_action('wsal_init', $this);
 		$s->Stop();
-		
-		// hide plugin
-		if($this->settings->IsIncognito())
-			add_action('admin_head', array($this, 'HidePlugin'));
 	}
 	
 	/**
@@ -237,7 +237,9 @@ class WpSecurityAuditLog {
 	 * Uninstall plugin.
 	 */
 	public function Uninstall(){
-		WSAL_DB_ActiveRecord::UninstallAll();
+		if ($this->GetGlobalOption("delete-data") == 1) {
+			WSAL_DB_ActiveRecord::UninstallAll();
+		}
 		wp_clear_scheduled_hook('wsal_cleanup');
 	}
 	
@@ -324,11 +326,7 @@ class WpSecurityAuditLog {
 	 * @internal To be called in admin header for hiding plugin form Plugins list.
 	 */
 	public function HidePlugin(){
-		$selectr = '.wp-list-table.plugins #';
-		$plugins = array('wp-security-audit-log');
-		foreach ($this->licensing->Plugins() as $plugin)
-			$plugins[] = strtolower(str_replace(' ', '-', $plugin['PluginData']['Name']));
-		?><style type="text/css"> <?php echo $selectr . implode(', ' . $selectr, $plugins); ?> { display: none; }</style><?php
+		?><style type="text/css">.wp-list-table.plugins #wp-security-audit-log { display: none; }</style><?php
 	}
 	
 	/**
